Added Database and rewrote APIs to work with the Database for fast retrieval of hashes.
